pyproject.toml
[project]
name = "deckboard-homeassistant"
version = "0.1.0"
description = "Home Assistant integration for the Deckboard Stream Deck UI library"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "deckboard",
    "hassapi>=0.3.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[tool.uv.sources]
deckboard = { git = "https://github.com/graphras-com/Deckboard", rev = "v0.0.1" }
